Move EncoderSimUtil to simulation package
