+++ hardware/ctre/src/main/java/io/github/frc5024/lib5k/hardware/ctre/sensors/TalonEncoder.java
@@ -1,6 +1,6 @@
 package io.github.frc5024.lib5k.hardware.ctre.sensors;
 
-import io.github.frc5024.lib5k.hardware.common.sensors.EncoderSimUtil;
+import io.github.frc5024.lib5k.hardware.common.sensors.simulation.EncoderSimUtil;
 import io.github.frc5024.lib5k.hardware.common.sensors.interfaces.EncoderSimulation;
 
 import com.ctre.phoenix.motorcontrol.can.BaseTalon;
